Made loginDlg fixed border and parameter text prompts that resize the dialog
Currently the prompts are exposed in the loginDlg so text can be passed
to each label
There is an exposed resize() javascript that can actually be places
inside loginDlg in an onVisible event so that it doesn't need to be
externally called.  I just left it external to call attention to it.

If you uncomment the parameters in the main.qml for loginDlg.  It will
revert to the demo long names I saved it as in  loginDlg.