streamlined the getUserList method and fixed the password validation code
streamlined the getUserList method and fixed the password validation
code
I didn't know how to test for the number of elements in userList_[] so I
hard coded it.

To send a SIGNAL you have to put the keywod "emit" in from of the signal
method.
The Dialog now works fully.